requirements.txt
click
requests
jinja2
pyyaml
validators
IPy
ansible
diagrams
ansible-vault
coloredlogs
colorama
